+++ xmtp_mls/src/identity/mod.rs
@@ -1,2 +1,1 @@
-pub mod v3;
 pub mod xmtp_id;
