+++ xmtp_mls/src/storage/mod.rs
@@ -4,7 +4,7 @@ mod serialization;
 pub mod sql_key_store;
 
 pub use encrypted_store::{
-    db_connection, group, group_intent, group_message, identity, identity_inbox, identity_update,
+    db_connection, group, group_intent, group_message, identity, identity_update,
     key_store_entry, refresh_state, EncryptedMessageStore, EncryptionKey, RawDbConnection,
     StorageOption,
 };
